+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # react-es6-webpack-skeleton ChangeLog
 
+## 2015-11-11
+### Changed
+* Upgraded to Babel 6
+* Use `module.exports` instead of `export default` in `webpack.config.babel.js`
+  for now (see [webpack #1578](https://github.com/webpack/webpack/pull/1578)).
+* Relax lint rule for "react/prop-types" to a warning instead of an error
+
 ## 2015-10-14
 ### Added
 * Initial release